Make CommandDispatcher simpler
