+++ commands/src/main/java/discord4j/commands/CommandDispatcher.java
@@ -15,63 +15,17 @@ import java.util.function.Predicate;
  *
  * @see CommandProvider
  */
+@FunctionalInterface
 public interface CommandDispatcher {
 
     /**
      * Called to handle the logic of invoking a command based on event context.
      *
      * @param event The event context.
+     * @param providers The {@link discord4j.commands.CommandProvider}s this dispatcher should consider for dispatching.
      * @param errorHandler The error handler to pass to commands.
      * @return A mono which emits executed commands.
      */
-    Mono<? extends BaseCommand> dispatch(MessageCreateEvent event, CommandErrorHandler errorHandler);
-
-    /**
-     * Gets the command providers registered with this dispatcher.
-     *
-     * @return The providers. It is expected that this returns an immutable copy of the internal backing set.
-     */
-    Set<CommandProvider> getCommandProviders();
-
-    /**
-     * Called to add a command provider dynamically.
-     *
-     * @param provider The provider to add.
-     * @return A {@link discord4j.commands.CommandDispatcher} instance for chaining, this can be a new instance or
-     * the current instance.
-     */
-    CommandDispatcher addCommandProvider(CommandProvider provider);
-
-    /**
-     * Called to add command providers dynamically.
-     *
-     * @param providers The providers to add.
-     * @return A {@link discord4j.commands.CommandDispatcher} instance for chaining, this can be a new instance or
-     * the current instance.
-     */
-    default CommandDispatcher addCommandProviders(Collection<? extends CommandProvider> providers) {
-        providers.forEach(this::addCommandProvider);
-        return this;
-    }
-
-    /**
-     * Called to drop a command provider dynamically.
-     *
-     * @param provider The provider to drop.
-     * @return A {@link discord4j.commands.CommandDispatcher} instance for chaining, this can be a new instance or
-     * the current instance.
-     */
-    CommandDispatcher removeCommandProvider(CommandProvider provider);
-
-    /**
-     * Called to drop command providers dynamically.
-     *
-     * @param providers The providers to drop.
-     * @return A {@link discord4j.commands.CommandDispatcher} instance for chaining, this can be a new instance or
-     * the current instance.
-     */
-    default CommandDispatcher removeCommandProviders(Collection<? extends CommandProvider> providers) {
-        providers.forEach(this::removeCommandProvider);
-        return this;
-    }
+    Mono<? extends BaseCommand> dispatch(MessageCreateEvent event, Set<CommandProvider> providers,
+                                         CommandErrorHandler errorHandler);
 }
